添加逻辑

+++ src/main/java/com/common/globals/server/impl/OnLinePlayerServer.java
@@ -59,7 +59,7 @@ public class OnLinePlayerServer implements IBaseServer{
 	 * @param playerId
 	 * @return
 	 */
-	public Player getPlayerById(int playerId) 
+	public Player getPlayerById(Long playerId) 
 	{
 		readLock.lock();
 		try
